+++ pages/about.tsx
@@ -1,13 +1,11 @@
+import { MdLayout, MdHeader, MdFooter } from "@/md";
 import { Box, Heading, Spacer, Text } from "@chakra-ui/react";
 import React from "react";
-import Layout from "../src/components/layouts/Layout";
-import LayoutFooter from "../src/components/layouts/LayoutFooter";
-import LayoutHeader from "../src/components/layouts/LayoutHeader";
 
 export default function AboutPage(): JSX.Element {
   return (
-    <Layout>
-      <LayoutHeader />
+    <MdLayout>
+      <MdHeader />
       <Box mt={20}>
         <Heading size="lg">About mediaDiary</Heading>
         <Box as="hr" my="2" />
@@ -17,7 +15,7 @@ export default function AboutPage(): JSX.Element {
         <Box as="hr" my="2" />
         <Text>lorem</Text>
       </Box>
-      <LayoutFooter />
-    </Layout>
+      <MdFooter />
+    </MdLayout>
   );
 }
