+++ src/components/Activity/config/fuego.ts
