+++ src/components/Activity/config/index.ts
@@ -1,0 +1,1 @@
+export { fuegoChartTop6, fuegoChartYear } from "./fuego";
