+++ src/components/Bookmarks/config/index.tsx
@@ -1,0 +1,1 @@
+export { fuegoBookmarkGet } from "./fuego";
