--- src/components/Bookmarks/fuego/index.tsx
@@ -1,1 +1,0 @@
-export { fuegoBookmarkGet } from "./fuegoBookmarks";
