+++ src/components/Day/config/fuego.tsx
