+++ src/components/Selected/config/fuego.tsx
