+++ src/components/Settings/config/index.ts
@@ -1,0 +1,1 @@
+export { default as useFuegoDelete } from "./useFuegoDelete";
